feature/add-more-data-into-database: Moved request, response and filter objects into specific folders for a better visualization. Refactored cost of living filter for University and max cost for Course. Added new script to insert more sample data into our database.
