+++ backend/src/main/java/com/lernia/auth/controller/AuthController.java
@@ -1,9 +1,9 @@
 package com.lernia.auth.controller;
 
-import com.lernia.auth.dto.RegisterRequest;
-import com.lernia.auth.dto.RegisterResponse;
-import com.lernia.auth.dto.LoginRequest;
-import com.lernia.auth.dto.LoginResponse;
+import com.lernia.auth.dto.request.RegisterRequest;
+import com.lernia.auth.dto.response.RegisterResponse;
+import com.lernia.auth.dto.request.LoginRequest;
+import com.lernia.auth.dto.response.LoginResponse;
 import com.lernia.auth.service.AuthService;
 import com.lernia.auth.entity.UserEntity;
 import com.lernia.auth.repository.UserRepository;
