+++ backend/src/main/java/com/lernia/auth/controller/UniversityController.java
@@ -2,7 +2,7 @@ package com.lernia.auth.controller;
 
 import com.lernia.auth.dto.UniversityDTO;
 import com.lernia.auth.dto.UniversityDTOLight;
-import com.lernia.auth.dto.UniversityFilter;
+import com.lernia.auth.dto.filter.UniversityFilter;
 import com.lernia.auth.service.UniversityService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
