+++ backend/src/main/java/com/lernia/auth/dto/request/ChangePasswordRequest.java
@@ -1,4 +1,4 @@
-package com.lernia.auth.dto;
+package com.lernia.auth.dto.request;
 
 import lombok.Data;
