+++ backend/src/main/java/com/lernia/auth/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.lernia.auth.dto;
+package com.lernia.auth.dto.request;
 
 public class LoginRequest {
     private String text;
