+++ backend/src/main/java/com/lernia/auth/dto/request/UserCourseRequest.java
@@ -1,4 +1,4 @@
-package com.lernia.auth.dto;
+package com.lernia.auth.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
