+++ backend/src/main/java/com/lernia/auth/dto/response/FavoritesResponse.java
@@ -1,5 +1,7 @@
-package com.lernia.auth.dto;
+package com.lernia.auth.dto.response;
 
+import com.lernia.auth.dto.CourseLightDTO;
+import com.lernia.auth.dto.UniversityDTOLight;
 import lombok.AllArgsConstructor;
 import lombok.Data;
