+++ backend/src/main/java/com/lernia/auth/dto/response/UpdateUserProfileResponse.java
@@ -1,11 +1,8 @@
-package com.lernia.auth.dto;
+package com.lernia.auth.dto.response;
 
-import com.lernia.auth.entity.UserCourseEntity;
 import com.lernia.auth.entity.enums.Gender;
 import lombok.Getter;
 import lombok.Setter;
-
-import java.util.List;
 
 @Getter
 @Setter
